+++ docs/changes.rst
@@ -19,6 +19,8 @@ v0.11.0 (in development)
 
 - Add support for the ``findadd`` command.
 
+- Add support for ``searchadd`` command added in MPD 0.17.
+
 
 v0.10.0 (2012-12-12)
 ====================